[refactor] split up handler classes
