+++ package.json
@@ -10,12 +10,14 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "jest": "^29.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.14.0",
     "sass": "^1.62.1"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.2",
     "@types/react": "^18.0.37",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.59.0",
